Add specific test rules for eslint.
As tests in mocha syntax tend to be more deeply nested (due to the
structure of describe etc.) this increases the maximal nesting
setting. Additionally this takes care of unused expressions (as
they are valid for chai e.g. in expect(foo).to.be.undefined) and
adds a rule that checks for e.g. it.only - the only should be
removed before tests are checked in to avoid that only these
tests run.